refactor(*): split gxi into multiple crates, use them as a workspace
