+++ src/editview/src/lib.rs
@@ -1,0 +1,9 @@
+#[macro_use]
+extern crate gxi_utils;
+
+pub mod edit_view;
+pub mod main_state;
+pub mod theme;
+
+pub use crate::edit_view::EditView;
+pub use crate::main_state::MainState;
