+++ src/editview/src/ui/close_tab.glade
