+++ src/editview/src/ui/find_replace.glade
