+++ src/gxi-config-storage/README.md
@@ -1,0 +1,11 @@
+# gxi-config-storage
+
+gxi's config backend, which uses a file based config for interacting with Xi and a dconf based backend for
+saving gxi specific config entries
+
+## Contributing
+
+Please see the docs on https://gxi.cogitri.dev/docs to learn more about gxi's inner workings. 
+[gtk-rs' site](https://gtk-rs.org/) offers documentation and examples about how gtk-rs works.
+
+Visit [Weblate](https://hosted.weblate.org/engage/gxi/) to translate gxi and its components.
