+++ src/gxi-config-storage/src/lib.rs
@@ -1,0 +1,7 @@
+#[macro_use]
+extern crate gxi_utils;
+
+pub mod errors;
+pub mod pref_storage;
+
+pub use crate::pref_storage::{Config, XiConfig};
