+++ src/gxi-linecache/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "gxi-linecache"
+version = "0.1.0"
+authors = ["Rasmus Thomsen <oss@cogitri.dev>"]
+edition = "2018"
+description = "gxi's linecache, used to cache Lines received from Xi-Editor"
+keywords = ["gxi", "xi"]
+categories = ["caching"]
+license = "MIT"
+homepage = "https://gxi.cogitri.dev"
+repository = "https://github.com/Cogitri/gxi"
+
+[dependencies]
+log = "0.4"
+serde = "1"
+serde_json = "1"
+
+[[bench]]
+name = "update"
+harness = false
