+++ src/gxi-linecache/README.md
@@ -1,0 +1,10 @@
+# gxi-linecache
+
+gxi's linecache which caches lines received from Xi via RPC for faster access
+
+## Contributing
+
+Please see the docs on https://gxi.cogitri.dev/docs to learn more about gxi's inner workings. 
+[gtk-rs' site](https://gtk-rs.org/) offers documentation and examples about how gtk-rs works.
+
+Visit [Weblate](https://hosted.weblate.org/engage/gxi/) to translate gxi and its components.
