+++ src/gxi-linecache/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod linecache;
+
+pub use crate::linecache::*;
