+++ src/gxi-peer/src/errors.rs
@@ -1,0 +1,5 @@
+#[derive(Clone, Debug)]
+pub struct ErrorMsg {
+    pub msg: String,
+    pub fatal: bool,
+}
