+++ src/gxi-utils/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "gxi-utils"
+version = "0.1.0"
+authors = ["Rasmus Thomsen <oss@cogitri.dev>"]
+edition = "2018"
+homepage = "https://gxi.cogitri.dev"
+repository = "https://github.com/Cogitri/gxi"
+keywords = ["gxi", "xi"]
+description = "Misc utility functions used throughout gxi"
+license = "MIT"
