+++ src/gxi-utils/src/globals.rs
