+++ src/gxi-utils/src/macros.rs
@@ -1,0 +1,24 @@
+#[macro_export]
+macro_rules! clone {
+    (@param _) => ( _ );
+    (@param $x:ident) => ( $x );
+    ($($n:ident),+ => move || $body:expr) => (
+        {
+            $( let $n = $n.clone(); )+
+            move || $body
+        }
+    );
+    ($($n:ident),+ => move |$($p:tt),+| $body:expr) => (
+        {
+            $( let $n = $n.clone(); )+
+            move |$(clone!(@param $p),)+| $body
+        }
+    );
+}
+
+#[macro_export]
+macro_rules! app_id {
+    () => {
+        gxi_utils::globals::APP_ID.unwrap_or("com.github.Cogitri.gxi.devel")
+    };
+}
