+++ src/gxi/src/ui/prefs_win.glade
